// src/generated/types.ts
export namespace AccountFields {
    export type id = number;
    export type address = string;

}

export interface Account {
    id: AccountFields.id;
    address: AccountFields.address;

}

export namespace AssetFields {
    export type id = number;
    export type account = string;
    export type assetId = string;
    export type name = string;
    export type supply = string;
    export type precision = number;
    export type symbol = string;
    export type txHash = string;

}

export interface Asset {
    id: AssetFields.id;
    account: AssetFields.account;
    assetId: AssetFields.assetId;
    name: AssetFields.name;
    supply: AssetFields.supply;
    precision: AssetFields.precision;
    symbol: AssetFields.symbol;
    txHash: AssetFields.txHash;

}

export namespace BalanceFields {
    export type id = number;
    export type address = string;
    export type assetId = string;
    export type balance = string;

}

export interface Balance {
    id: BalanceFields.id;
    address: BalanceFields.address;
    assetId: BalanceFields.assetId;
    balance: BalanceFields.balance;

}
// ...
    export type timestamp = string;
// ...
    export type block = number;
// ...
export namespace TransferFields {
    export type id = number;
    export type asset = string;
    export type from = string;
    export type to = string;
    export type txHash = string;
    export type value = string;
    export type amount = string;
    export type block = number;
    export type timestamp = string;

}

export interface Transfer {
    id: TransferFields.id;
    asset: TransferFields.asset;
    from: TransferFields.from;
    to: TransferFields.to;
    txHash: TransferFields.txHash;
    value: TransferFields.value;
    amount: TransferFields.amount;
    block: TransferFields.block;
    timestamp: TransferFields.timestamp;

}